feat: refactor participant API routes to use ParticipantsController for improved structure and maintainability
